backend/requirements.txt
fastapi
uvicorn[standard]
celery
redis
python-multipart
sqlalchemy
alembic
passlib[bcrypt]
pyjwt[crypto]
python-dotenv
